refactor: rename package to auditlog_fastapi and update imports
