+++ auditlog_fastapi/db/beanie_document.py
