+++ auditlog_fastapi/db/sqlalchemy_table.py
