+++ auditlog_fastapi/exceptions.py
